Fix Android BuildStep bug
We use the output of ADB for a number of things in our BuildSteps. When ADB
runs for the first time after a reboot, it prints out additional lines
indicating that the ADB daemon is starting. In this one case, we only care
about the last line of output, whether that message is printed or not.

This was causing failures on Nexus 10.
Review URL: https://codereview.appspot.com/6849047

git-svn-id: http://skia.googlecode.com/svn/buildbot@6402 2bbb7eff-a529-9590-31e7-b0007b416f81