Change build status behavior for skipped steps
r6336 modified the behavior so that, when we skip steps on builds for
rebaseline commits, the skipped steps inherit the result of the last time they
ran. While this is correct in most cases, there is a race condition in which
multiple slaves running the same builder produce incorrect results if the
status changes just before a rebaseline commit so that the newly-successful or
failing build is started before but finishes after the build for the rebaseline
commit. The rebaseline build will inherit the wrong statuses.

Instead, we abandon the notion of inheriting statuses and just mark the build
as "some steps skipped" if all of the steps which ran succeeded but some were
skipped. If some steps were skipped but one or more which ran failed, we still
mark the build as a failure. This eliminates the confusion, but I'm not sure
how it will affect breakage notifications since "some steps skipped" still
counts as a success.
Review URL: https://codereview.appspot.com/6843055

git-svn-id: http://skia.googlecode.com/svn/buildbot@6456 2bbb7eff-a529-9590-31e7-b0007b416f81