Allow exceptions to propagate in merge_into_svn
We were catching exceptions in _SetProperty in case there were no files with
the requested extension, but since we now use _FindFiles (which returns an
empty list in this case) this is no longer required.

This was hiding other problems while uploading GM results.
Review URL: https://codereview.appspot.com/6855105

git-svn-id: http://skia.googlecode.com/svn/buildbot@6562 2bbb7eff-a529-9590-31e7-b0007b416f81