Remove SkiaBuild
SkiaBuild overrides Build in order to force a build with skipped steps to have
SKIPPED as its result.  This was originally done to prevent confusion resulting
from a failed build appearing to be fixed when in reality the failing step was
skipped.  This turns out to be a rare occurrence, however, and SKIPPED does not
play well with the console page.

Addresses https://code.google.com/p/skia/issues/detail?id=1190

(SkipBuildbotRuns)
Requires a master restart.
Review URL: https://codereview.chromium.org/13533003

git-svn-id: http://skia.googlecode.com/svn/buildbot@8508 2bbb7eff-a529-9590-31e7-b0007b416f81