If a compile fails, clean and try again
Note that this does not affect the failure/success of the build as a whole (if
any compile fails, the whole build is a failure, even if the subsequent tests
run and succeed).

This behavior is not replicated on Android, because the Android bots already
clean before every build.

Addresses https://code.google.com/p/skia/issues/detail?id=1286

(SkipBuildbotRuns)
Requires a master restart.

R=rmistry@google.com

Review URL: https://codereview.chromium.org/15995018

git-svn-id: http://skia.googlecode.com/svn/buildbot@9424 2bbb7eff-a529-9590-31e7-b0007b416f81