Run skimage on the bots.
Run our image decoders on a set of images for regression testing.

On each run, we will download a set of expectations from a git repository,
and download the images from Google Storage. Both are copied to the
device (if necessary), and then run the tests.

TODO: Add a script to perform the rebaseline.
FIXME: The function to determine the correct folder is incorrect.

BUG=https://code.google.com/p/skia/issues/detail?id=1241
R=borenet@google.com

Review URL: https://codereview.chromium.org/16226005

git-svn-id: http://skia.googlecode.com/svn/buildbot@9771 2bbb7eff-a529-9590-31e7-b0007b416f81