Pass folder name to skimage.
On Android, passing each file individually results in a command
line that is too long. The reason for passing each file individually
was to avoid running the test on the timestamp, which reports a
failure. Since the timestamp is not present on platforms where
passing each file individually causes a problem, just pass the
directory in that case.

Fixes Android builds.

R=borenet@google.com

Review URL: https://codereview.chromium.org/18210002

git-svn-id: http://skia.googlecode.com/svn/buildbot@9808 2bbb7eff-a529-9590-31e7-b0007b416f81