Revert "Raise a warning when failing run_decoding_tests."
This reverts commit fd301bdda0e8f2fe6232ddfb613c851f6560dd2b.

This was intended to be a temporary commit while I added some new
files. I need to come up with a better way to add new files without
turning the bots red... For now, reverting this, so warning goes
back to being an error.

TBR=borenet@google.com

Review URL: https://codereview.chromium.org/18854012

git-svn-id: http://skia.googlecode.com/svn/buildbot@9917 2bbb7eff-a529-9590-31e7-b0007b416f81