Updates to skimage expectations.
BUG=skia:1399

Use the builder name as the expectations filename for both
reading and writing. A separate CL in trunk will add the new
files to read. Raise a warning if there is no expectation
file, to ensure that we don't skip checking in any
expectations files.

BUG=skia:1555

Only push one set of expectations to the device: the one
corresponding to the builder. This removes the trybot suffix
if necessary, so the right file will be used either way.

Modify GetWaterfallBot to return the name of the
corresponding waterfall bot whether the builder_name is a
waterfall bot or its trybot. Update the other caller of
GetWaterfallBot.

R=borenet@google.com

Review URL: https://codereview.chromium.org/22640011

git-svn-id: http://skia.googlecode.com/svn/buildbot@10921 2bbb7eff-a529-9590-31e7-b0007b416f81