Write skimage results to full builder_name.
UploadSKImageResults attempts to upload a folder matching the full
builder_name, including _Trybot if the builder is a trybot. (This
allows us to see the results of a try.) Fix a bug where
run_decoding_tests wrote the results to a folder matching the
corresponding waterfall bot (resulting in the upload step not finding
anything to upload).

This fixes a bug where all trybot runs fail UploadSKImageResults.

R=borenet@google.com

Review URL: https://codereview.chromium.org/26824003

git-svn-id: http://skia.googlecode.com/svn/buildbot@11729 2bbb7eff-a529-9590-31e7-b0007b416f81