Add SourceStamp status page
This displays build statuses for a given source stamp, which corresponds
to either a "force build" or a new code revision. The current version
displays a table containing all of the buildsets (Tasks) which have run
for the given source stamp. In the future, I will display these with a
dependency graph.

R=epoger@google.com

Review URL: https://codereview.chromium.org/27354003

git-svn-id: http://skia.googlecode.com/svn/buildbot@11784 2bbb7eff-a529-9590-31e7-b0007b416f81