Write skimage output to checksum filenames.
Set a default pref config, which will be a part of the
resulting filename.

Put the output in a separate directory from actuals,
and perform a merge in the upload, so that files will
now be stored according to hash rather than builder.

Depends on https://codereview.chromium.org/26297004/

BUG=skia:1466
R=epoger@google.com

Review URL: https://codereview.chromium.org/27389002

git-svn-id: http://skia.googlecode.com/svn/buildbot@11901 2bbb7eff-a529-9590-31e7-b0007b416f81