Fix the GateKeeper to do timestamp comparisions instead of rev comparisions.
Subclasses gatekeeper.GateKeeper and overrides the isInterestingStep method. Saves the start time of the build that last closed the tree and compares it to the current build to determine if the current build is older and the tree should not be closed.

BUG=skia:1866

(SkipBuildbotRuns) Requires a master restart.

R=borenet@google.com

Review URL: https://codereview.chromium.org/108393002

git-svn-id: http://skia.googlecode.com/svn/buildbot@12556 2bbb7eff-a529-9590-31e7-b0007b416f81