Commenting out timestamp comparison logic in Skia gatekeeper.
The gatekeeper was not working with the timestamp comparison because when the Retry_BuildGm step reaches this point it thinks that this build is different than the BuildGm build (it was actually the same build) and proceeds to not close the tree when it should close it.

With this change older builds may reclose the tree but atleast the gatekeeper will be working while I investigate a way to use timestamps for only the 'Retry_' steps.

Review URL: https://codereview.chromium.org/122303002

git-svn-id: http://skia.googlecode.com/svn/buildbot@12854 2bbb7eff-a529-9590-31e7-b0007b416f81