Fix web status page reloading
The javascript_revision was getting set to the revision that the master was
running, rather than the most recently checked out revision as it should have
been.  As a result, if the master gets synced without a restart, the web status
pages will refresh over and over instead of just once.

Requires a master restart.

BUG=
R=jcgregorio@google.com

Review URL: https://codereview.chromium.org/131803004

git-svn-id: http://skia.googlecode.com/svn/buildbot@12986 2bbb7eff-a529-9590-31e7-b0007b416f81