Reorganize master_builders_cfg to define compile and test/perf bots separately
This will allow us to put associated compile and test/perf bots on different masters.  It's a big change, but fortunately the factory_config test demonstrates that none of the builders themselves have actually changed, with the exception of the Logan bot, whose name did not adhere to the naming scheme of the others and needed to change to make this work.

BUG=
R=epoger@google.com

Review URL: https://codereview.chromium.org/139293006