Stop trying to use latest git features.
The canary bot only has git 1.7, which does not support 'show -q'.
The '-q' is not *needed*; I only want a small piece of the commit
message anyway, and with or without the '-q' the rest of it gets
removed later on when I do a substring.

R=borenet@google.com, epoger@google.com
TBR=borenet@google.com, epoger@google.com

BUG=skia:2409

Author: scroggo@google.com

Review URL: https://codereview.chromium.org/239243015