Fail -> warn on repeat merge. Remove SVN rev #.
When attempting to merge Skia ToT into master-skia, do not report
failure if ToT has already been merged. Instead, raise a BuildStepWarning.

I originally reported it a failure assuming that it would rarely
happen, and if it did, it was due to some sort of problem. In
practice, due to timing issues, it happens fairly regularly, and it
has not been a sign that there was any sort of problem.

Do not use the old SVN revision number, which soon will no longer exist.
Instead, use the commit hash plus date of commit.

R=borenet@google.com, djsollen@google.com

Author: scroggo@google.com

Review URL: https://codereview.chromium.org/284233003