Delete historical bench data on buildbots
I think this should be fine, since all of the data is in GS now. This change
prevents us from always uploading all of the data on every build, which is
expensive in the normal case and is a big problem in the case of trybots.

BUG=skia:2225
R=bensong@google.com, epoger@google.com

Author: borenet@google.com

Review URL: https://codereview.chromium.org/305623002