Add flag to control how much data is loaded.
Additionally moved over to glog for logging, as opposed to the native logging package.

Finally fixed a bug where we would crash if no data was generated for a day.

Note that after this CL lands you should add

 --alsologtostderr

to the command line when running locally.

BUG=skia:
R=mtklein@google.com, kelvinly@google.com, bensong@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/341073002