This CL adds a new /clusters/ page that shows the results of the k-means clustering on all the Skp measurements.
For each graph we show the centroid that was calculated, the total number of traces that fell into that cluster, and also show how the Params vary for each cluster by varying the font size.

This also fixes a bug in the mutex handling for refreshing the data.

BUG=skia:
R=mtklein@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/341193005