Save and restore state from gzipped JSON files.
This is step 2 on the way to moving to 'tiles'. Note that
this CL also removes the --full_data flag. I will put up
an skps.gz and micro.gz file on Google Storage that anyone can copy to /tmp/tiles for running locally and thus avoid
the long startup the first time.

BUG=skia:
R=mtklein@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/332093009