Add the ability to have more than one config file.
Add a bodytest parameter to the prober config that selects a function to run over the response body.

Add a bodytest function that tests that buildslaves all report as connected.

BUG=skia:2697

I also had started on a Go program to read the global config and write the buildbots.json file, but then realized that the port number was stored separately from the IP address of the buildbot, and then I got sad, and tired. I can put up that code in a separate CL if someone want to take a crack as finishing it.

R=borenet@google.com, mtklein@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/345703004