Updates the repository so that it can be used as a normal Go build target.
The vast majority of the changes are just moving the files
around, the directory structure has been flattened so that
we run out of perf/ and the source files have been moved
from /perf/server/src/<foo> to /perf/go/<foo>. In addition
the references to the local libraries are now scoped
properly, so instead of

  import "types"

They are now scoped as:

  import "skia.googlesource.com/buildbot.git/perf/go/types"

which avoids conflicts with the standard library.

The only real code change is to perf.go, which used to set
its working directory to the executable's directory, is now
setting its working directory to two directories up from the
perf.go source file, so that no matter where it's run from
it will find the JS and template sources.

Also, now that all executables are now going into
$GOPATH/bin, the .gitignore can be cleaned up.

See the changes to README.md and Makefile to see how this changes development.

BUG=skia:
R=mtklein@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/436633002