Disable CPU benching on GN, and CPU testing on GN, N7, and N10.
We worked out that these configurations are just unnecessary load,
covered just as well by devices with faster variants of the GPU or CPU.

Not bothering with tools other than DM and nanobench.

As a bonus, let's try running DM with --gpu enabled on the GN again.
Otherwise, it won't be running anything and might feel embarrassed.

BUG=skia:
R=djsollen@google.com, bsalomon@google.com, mtklein@google.com

Author: mtklein@chromium.org

Review URL: https://codereview.chromium.org/524723002