Add trybot data to the ingester.
Kept separate from normal build data because it is ingested at a faster rate, and stored differently, that is, the trybot data is store in the database and not in Tiles.

Changes include breaking out the last updated timestamp reading/writing code so it can be used by both types of ingestion.

Also make ingester.GetFilesFromGSDir and ingester.GetBenchFiles public so they can be called from trybot.

BUG=skia:
R=mtklein@google.com

Review URL: https://codereview.chromium.org/524523004