Move to a beefier instance for skiaperf.
Moving from 1 core, 3.7 GB RAM to 8 cores, 52GB RAM.

This is a new instance with a different name, once skiaperf is up and running on it I will update the DNS records to point to the new instance.

Also making this a beefier instance since this is where we'll run gold.

Once gold is up and running we can shut down the old RBS and free up 8 cores.

BUG=skia:
R=rmistry@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/549373004