PTAL
I am adding DB versioning to the dabase. I tried a third party tool (goose), but that did not work for what we needed. So I implemented a very basic approach that stores
the version in the data base.

There is still an command missing (i.e. 'migratedb') that will prompt for the root password and do the actual migration.

Btw, this should be generic enough to work for other apps where we need a DB.

BUG=skia:
R=jcgregorio@google.com, rmistry@google.com, bensong@google.com

Author: stephana@google.com

Review URL: https://codereview.chromium.org/608273002