repair the webtry setup process
The previous version of the script didn't properly update the code for the webtry
user, so only changes to skia would really be pulled, not changes to the actual
webtry server itself.

This version of the setup script is 100% automated and brings everything up to date.

It's also a little more verbose about what it's doing, so it's easier to debug.

BUG=skia:
R=tfarina@chromium.org
TBR=jcgregorio, mtklein, tfarina

Review URL: https://codereview.chromium.org/636173003