webtry: Simplify template construction.
text/template and html/template both have a helper function called
"Must" that according to the docs in http://golang.org/pkg/text/template/#Must
and http://golang.org/pkg/html/template/#Must, wraps a call to a function
returning (*Template, error) and panics if the error is non-nil.

So we don't need to that ourselves. Which is a good thing and makes code
cleaner.

BUG=None
TEST=None
R=jcgregorio@google.com

Review URL: https://codereview.chromium.org/642243004