webtry: Add logserver to serve /tmp/glog.
This setup and runs a logserver, like the one being used by skiaperf, to
serve the directory /tmp/glog, so we don't need to ssh into the instance
to see the logs of the webtry application.

This is a follow up to suggestion made in codereview.chromium.org/655323002.

BUG=None
R=jcgregorio@google.com

Review URL: https://codereview.chromium.org/643903003