Redo the login-sk element in Polymer.
Add the platform.js into our minified set of JS we include to begin with.

Then add in the polymer.html HTML Import to headers.html template. In theory we could add the contents of polymer.js to our minified JS also, but that would break the ability to use all the other core- and paper- components, which assume they have a relative location to polymer.html.

Finally, tweak the CSS to add a judicious use of ::shadow so that our style doesn't change.

TBR=borenet
BUG=skia:

Review URL: https://codereview.chromium.org/643893004