get rid of the silly javascript 'safe-image' load things
Prior to this, I was hiding the image wrappers and showing them if
the HTTP request succeeded, but this was silly; the PDF just checks
for file existence on the server and builds the template properly.

Now the images do the same, so all that crummy javascript polluting
the global namespace can disappear.

BUG=skia:
R=jcgregorio@google.com

Review URL: https://codereview.chromium.org/700213006