Polymerize Query, which handles all the selections across params for Traces.
The query control is used on the front page, the clustering page, and the compare page, and all of those pages have been updated to use the <query-sk> element.

<toggle-display-sk> has been moved out into its own .html file.

On both the cluster and combine pages I moved the JS from /res/js/ back into the main page, with the idea of keeping the markup and logic in the same page. This will also help if we decide to turn each of those pages into elements.

CSS for the main page is sprinkled between templates/index.html and res/css/perf.css, which I will fix in a follow-up CL.

BUG=skia:

Review URL: https://codereview.chromium.org/695773006