Ignore bug_chomper's downloaded oauth_client_secret json file.
When running bug_chomper's ./run_server.sh script, it will download
oauth_client_secret.json from Google Storage so it can logging in
code.google.com. Since that file is not checked in the the source tree,
lets not show it in the 'git status', which is annoying.

BUG=None
R=borenet@google.com

Review URL: https://codereview.chromium.org/726573003