Fix some sk Promises.
First fix is to turn the domReady() function into DomReady an instance
of a Promise, which is simpler and doesn't require building a new
Promise each time.

Also add a WebComponentsReady which does the same thing for the 'polymer-ready' event
fired by Polymer. This improves the situation on FireFox (but doesn't fix us completely).

BUG=skia:

Review URL: https://codereview.chromium.org/759813002