Cleanup Error() usage.
As stated in golang.org/pkg/fmt:

3. If an operand implements the error interface, the Error method will
be invoked to convert the object to a string, which will then be
formatted as required by the verb (if any).

So it seems redundant to call err.Error() on those functions, since that
will be done for us for free.

This was noted in https://codereview.chromium.org/787133002/

BUG=None
R=jcgregorio@google.com

Review URL: https://codereview.chromium.org/786953004