Speed up doc server.
In the process of speed testing docserver and the skfe configuration I discovered that by default all processes on GCE instances have a default ulimit of 1024 open file handles, which isn't enough for a web server under heavy load.

This CL also adds the code for the load tester, and adds some caching for the Markdown documents.

BUG=skia:

Review URL: https://codereview.chromium.org/834273003