This CL makes it easy to test Skia gold against live data.
It extends the tiletool to dump a (subset of a) tile to JSON and adds code to parse the JSON back into a Tile.
It adds a a basic test against live data from the current tile and fixes a bug (in util.IntersectIntSets) that caused wrong query results.

BUG=skia:2930

Review URL: https://codereview.chromium.org/842313002