Fast path image comparison. If two images are the same size and they can both be converted to NRGBA then do an image comparison via their Pix members. This method of comparing images is 15x faster than the old method.
Note that the old method of comparing images is still kept in place for comparisons that can't use the fast path.

The fast path comparison also seems to the fix the bug seen on triage day where 565 images are seen to differ 100% from 8888 images solely based on the alpha channel.

BUG=skia:

Review URL: https://codereview.chromium.org/828623005