This adds caching to the analyzer to make startup much faster by using the latest known state.
Then it calculates the diffs between all digests (per test) in the background and updates the state to the latest tile data.

It moves writting diff files and creating the thumbnails
into the FileDiffStore and switches to relative filenames
for diff images and their thumbs.

The later is a requirement for moving data around and pre-caculating diff values apriori.

BUG=skia:

Review URL: https://codereview.chromium.org/899433007