Adds and Flag for CPU profiling.
It disables thumbnailing because the resize calls were responsible of 50% of the CPU time spent, making diffing twice as slow (in terms of CPU).

The same is true for compression of the output images.
We can re-enable these as we do more testing.

Adds some more parallelism.

BUG=skia:

Review URL: https://codereview.chromium.org/898793002