This adds a build_release script to create a Debian package for skia gold.
It extends the releash.sh with the BYPASS_UPLOAD flag for testing.

It also adds basic file concatenation for the Gold app, to reduce the number of files that need to be included into the package. This is just meant to be enough to work for be able to deploy the package.

A future CL will harness wbtools/webtools.mk.
BUG=skia:

Review URL: https://codereview.chromium.org/886353007