Status: Separate builds from commits in cache
Ensures that we only have a single copy of a given build in play at any
time.

This is leading up to the addition of comments to builds; I need to be able to get a buildbot.Build object from the cache by ID, add a comment, and update it in the database.  The old cache made it difficult to determine where to update builds, so they ALL would've needed to be reloaded whenever a comment was added.

BUG=skia:3125

Review URL: https://codereview.chromium.org/918623003