A new diff color scheme that shows how different pixels are.
Here is the new scale:

  https://screenshot.googleplex.com/vEJq4Wc7sO.png

Here is an example diff that includes both RGBA and Alpha only differences:

  https://screenshot.googleplex.com/3raAmYNOU2.png

Here is a zoomed view on that diff:

  https://screenshot.googleplex.com/outnqMs9ak.png

BUG=skia:

DO NOT SUBMIT until we're ready to rebuild the entire cache
of diff images.

CL also adds a help page explaining the diff colors.

Review URL: https://codereview.chromium.org/925013002