Display reverted CLs and their reverters.
Summary of changes:
* Reverted CLs will be displayed with a red undo icon.
* Clicking on the red undo icon will highlight the reverted commit and will also open its commit popup.

In the future we can extend this to instead draw a line from reverted to reverter if needed.

The change is staged at http://piraeus.cnc.corp.google.com:8000/
The layout of branches and commits looks broken but I do not think it is because of my change, I think its because of incompletely loaded commits.  I see a lot of strange layout when the builds first start loading. Clicking on 'All' or 'failures' fixes this layout.

If you look closely the new icon shows up within the builders column. I do not know why this happens, I added logic similar to what "commentIndicator" does. Any ideas?

BUG=

Review URL: https://codereview.chromium.org/975833002