This is the first of several CLs to replace the current analyzer package.
It introduces the kernel module with the Storage and Kernel structs that serve as container for storage and functionality. This will make it easier to pass around
the ever increasing number of storage objects.

This implements the blame list feature by solely processing a tile from the tile store. It adds the DigestStore interface that aggregates information about individual digests (implementation forthcoming).

BUG=skia:

Review URL: https://codereview.chromium.org/963923002