Make AlertServer store alerts in a database
This is a major rewrite of the AlertServer guts which will result in
the server storing every alert and comment for posterity. This fixes
skia:3290 ("Alert Server: persist alerts through server restarts") and
will allow us to do interesting analyses on the database, like which
alerts fire most often. The reorganization will also enable us to add
an endpoint for POSTing alerts to the database, so that we can fire
arbitrary alerts from anywhere.

BUG=skia:3290

Review URL: https://codereview.chromium.org/983143005