Adding a head toggle to /2/ summary.
Note that adding support for head on the grid interface will be in a second CL and at that time links from /2/ to the grid will preserve the head value.

Also disable calculating the diameter for now since it is mostly useless and dramatically slows down everything.

BUG=skia:

Review URL: https://codereview.chromium.org/1054313002