Speeding up CalcSummaries.
The surprising thing here is that the slowest part of CalcSummaries was sorting the traces.

The only reason I was sorting the traces was to cluster the traces together by test name, but placing them in a map keyed on test name removes the need for that sort.

I also switched Summary to calculate and store summaries for head, which is now the default on /2/, so this speeds up loading that data, going from about 5s down to 5ms. Loading non-head summaries has sped up from 5s down to 1.5s.

BUG=skia:
TBR=stephana

Review URL: https://codereview.chromium.org/1072443002