Buildbot dashboard: also display avg time in step across all builds
Make the "Step Times" chart a double-bar chart which displays both the
average duration of the step for builds where that step ran (what we
currently have) AND the average time spent in the step across all
builds, including those which did not run the step (the duration of
the step in those builds is zero).

This makes it more clear which steps are "occasionally bad", for
example, DownloadSKPs does take a long time, but it only runs about
once a week per bot, so the average across all builds is pretty low.

BUG=skia:3603

Review URL: https://codereview.chromium.org/1052553005