Add triage status to the dots legend.
Rename OtherHashes in PolyDetailsGUI to be OtherDigests, to fix the nomenclature, also change the value from a slice of digests to a slice of DigestInfo's which contain the digest and its status.

Most of the rest of the change is flowing that data into the UI.

The last little bit is adding the tricon-sk element which displays the triage status as a small icon.

Screenshot:

  https://screenshot.googleplex.com/tQrzAuJoaP.png

BUG=skia:

Review URL: https://codereview.chromium.org/1094793003