global: Don't try to write an error message to an io.Writer that may be in an error state.
So I found this in the skia-gold logs, that sometimes nginx disconnects from the server
before getting the complete response (which is something we should also look into) but
when that happens the handler may call util.ReportError, which then tries to write
a 500 response to that already broken socket, either "connection reset by peer",
or "broken pipe", so writing to that socket just generates more errors. I found thousands
of log lines in skia-gold logs with this error.

This CL changes every call to util.ReportError that is used for errors
reported when writing to a socket and replaces them with glog.Errorf.

LGTM for the files you feel comfortable reviewing.

BUG=skia:

Review URL: https://codereview.chromium.org/1168503005