gold: Break out blames into their own element. Update testsummarydetails to use the new element.
This CL has no user visible affect, the UI remains the same.

The highlighting is now handled through an event. The blame-sk emits the event
and testsummarydetails receives that event and calls the new highlight method
of the dots-sk element.

The first in a series of CLs to refactor some of the UI into smaller orthogonal pieces.
Next CL will break out the dots legend as its own element, followed by a
CL that uses Mailbox to distribute the commit into to all the elements
that need it, which should reduce our redundant HTTP traffic.

BUG=skia:

Review URL: https://codereview.chromium.org/1191253006