gold: Rebuild search and results.
The actual response from /search2 comes back in around 300ms. The rest of time displaying a long page is Polymer 0.5 slowness, which should be fixed once we move to Polymer 1.0.

Also, given how fast the list is to produce and sort we can now add pagination to list results, making the display much faster, and removing the total limit on search results.

The /search2 endpoint will remain hidden for now, moving to replace /search and /blamelist in a later CL.

BUG=skia:
TBR=mtklein

Review URL: https://codereview.chromium.org/1215833003