Delete fix_archives application from CT.
I initially wrote this to try to automatically detect and delete flaky webpage archives from the CT repositories. However, it is not being used since I am no longer convinced that the time it takes to run and maintain this is worth the limited gains from running it.

BUG=skia:

Review URL: https://codereview.chromium.org/1243083002