Add NSQ to GrandCentral
This adds a script to package nsq (see http://nsq.io) to make it
pushable. This is a minimal installation to use NSQ as a message transport and to be able to test  https://codereview.chromium.org/1212123004/

It also fixes a bug in the push server.

BUG=skia:
R=borenet

Review URL: https://codereview.chromium.org/1251873002