Add support for scheduling repeated runs in the frontend and in DB.
Summary of changes in this CL:
* Adds a new repeat-after-days polymer element to all task scheduling pages.
* Displays if the task is set to repeat on all task run history pages.
* Displays tasks which are scheduled for the future in the 'View Queue' page.
* Makes the necessary modifications to tables and task schedulers to make the above work.

What is not done yet is completing the implementation of the goroutine which will actually schedule the tasks. I left this incomplete because Ben's out for review CL https://codereview.chromium.org/1257893004/ might change how to update tasks.

BUG=skia:4026

Review URL: https://codereview.chromium.org/1261883002