gold: Making paramset-sk handle displaying multiple paramsets.
A later CL will roll this into digest-details-sk to display both the paramsets
for the digest and its closest negative or positive digest.

Functionally this change is a noop, nothing in the UI changes in the current UI, the ability to display multiple paramsets and be non-clickable will only show up in digest-details-sk.

BUG=skia:

Review URL: https://codereview.chromium.org/1270733002