Build Scheduler: multi-buildslave support
Before this change, the queue for each builder represented the best,
second best, third best, etc, build candidate *right now*. In other
words, we only really cared about the candidate at the head of the
queue. This means that, if we have multiple slaves running one builder,
the second slave may choose a sub-optimal build to run if the queue
isn't updated between the first slave choosing a build and the second
slave choosing a build.

After this change, the queue for each builder is actually a queue: the
build candidate at the head of the queue is the optimal choice to build.
When that is popped from the queue, the next candidate is also the
optimal one, and so on.

BUG=skia:4194

Review URL: https://codereview.chromium.org/1290243002