Add datahopper_internal, which ingests builds from tradefed, obfuscates the target names, and pushes the data into the buildbot database.
Also serves an auth protected endpoint that redirects from the obfuscated target names to the internal site.

Screenshot of the redirect landing page:

   https://screenshot.googleplex.com/28KXVjYJWCN.png

   That is the landing page for the URI: /t/purple-paper-958871/12/

See the included db_example.txt for the state of the database after running for a little bit. It will eventually run at https://internal.skia.org.

The setup for the instance that this will run on will be done in a separate CL.

BUG=skia:4111

Review URL: https://codereview.chromium.org/1309263002