gold: Fix for position: fixed of dialog and floating action button on ignores edit page.
In Polymer 0.5 core-scaffold has a transform: translate CSS rule, which breaks what you
would expect the normal behavior of position: fixed to be (but note that this is correct
according to the spec). The fix is to move the dialog and fab out of the scaffolding.

  https://groups.google.com/forum/#!msg/polymer-dev/42DTbwDahN4/C73AKRozkOgJ

Note that this is not a problem with the iron-* elements from Polymer 0.9.

BUG=skia:

Review URL: https://codereview.chromium.org/1313573004