[CT] Display description of SKP repository on the Lua scripts page
Motivation:
For skia:4336 I created a "special" chromium build which will be used to capture SKPs with no font data. I am going to mention this in the description of the SKP repository and would like the information to show up on the Lua Scripts page.

BUG=skia:4336

Review URL: https://codereview.chromium.org/1342363003