This implements a gRPC-based frontend server for BoltDB. It allows us to run an embedded database on the instances that can be shared by multiple processes.
This is the first version. It basic tests, but will require more testing. It currently does not support range scans and prefix scans. These features and the tests will be added in a subsequent CL.

The first application to use this is the ingestion package for trybot data.

BUG=skia:2347

Review URL: https://codereview.chromium.org/1341423007