[CT] Remove support for draw_properties from CT
Motivation:
On runs of draw_property only 3178 webpages out of ~9k report results. This is likely because the majority of the web pages in the alexa set cannot scroll (because they are home pages and not deep links) and thus have fewer samples for draw properties to work with.
Removing support because this benchmark does not seem to provide any useful data.

BUG=skia:
NOTRY=true

Review URL: https://codereview.chromium.org/1404543003