This implements a new ingestion framework. It's a rewrite of the infra/go/ingester package, that is significantly simpler, tested and with clear interfaces.
The framework also allows event driven ingestion which promises to significantly reduce latency between a result file being available and it showing up in an app.

The large number of changed files are a result of the introduction of vcsinfo which is a abstract interface of gitinfo. This allows easier testing and the use of a VCS other than git (though that was not a primary objective).

The next CL will add more tests and

BUG=skia:

Review URL: https://codereview.chromium.org/1419533009