tracedb: use fast calls to load data.
This isn't much faster on the first load, but the subsequent loads are much faster since the caches are filled.

On my desktop the first 50 commits tile load is 2m31s and the second load is 28s. Times will be faster in production since most of the time is spent waiting for the network. Here the total run to load a 50 commit tile twice took 3m, but we only burned 64.5s of CPU, the remaining time was waiting for the network. And 30s of the CPU was rebuilding the Tiles paramset, which I might be able to speed up in a subsequent CL.

(pprof) top20 -cum
22.55s of 100.55s total (22.43%)
Dropped 264 nodes (cum <= 0.50s)
Showing top 20 nodes out of 107 (cum >= 5.57s)
      flat  flat%   sum%        cum   cum%
     1.99s  1.98%  1.98%     64.51s 64.16%  runtime.duffcopy
     0.04s  0.04%  2.02%     37.93s 37.72%  unlockOSThread
     1.63s  1.62%  3.64%     30.07s 29.91%  go.skia.org/infra/go/tiling.GetParamSet
     0.57s  0.57%  4.21%     12.03s 11.96%  runtime.selectgoImpl
         0     0%  4.21%      9.42s  9.37%  main._main
         0     0%  4.21%      9.42s  9.37%  main.main

Once this is running in prod for a while we can go back and remove the GetValues call from the gRPC service.

BUG=skia:

Review URL: https://codereview.chromium.org/1516463003