Make fuzzer frontend aware of pending version changes.
This includes a refactor of the older version_updater code so there is less
duplication between frontend and backend.

When a pending version is detected, the frontend will check out the new skia version
and start building the AST.  When the current version changes, it re-executes the same
code as when it starts up to download the fuzzes.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1538333002