Add the ability to link to a single fuzz
There is some refactoring in this commit to names that make a little more
sense (gofmt -r for the win).  Additionally, the clickable icons all light up when moused over.

The search by name requires that all of the sub lists of binary and api
fuzzes be sorted, so that is added in as well.  Given that GCS gives us
fuzzes that are sorted most of the time, it makes sense to append the fuzz
to the existing list and check to see if that is the right place.  Most
of the time, this will be true.

One last fix was properly showing/hiding the expansion on the stacktrace
views.

BUG=skia:4438

Review URL: https://codereview.chromium.org/1559523002